requirements.txt
streamlit
streamlit-autorefresh
pillow
transformers
torch
sentencepiece